Fix internal hyperref names
As stated in the hyperref README[1], the count1to packages confuses the
internal numbering of the hyperref package. In the demo PDF, this can be seen
in the TOC: Clicking on the "4.1 Including code listings" TOC entry leads to
"3.1 A “meta” LaTeX style file". This problem also gives two warnings about
duplicate names.

Fix this by using an adaptation of the suggested fix from the hyperref README.

[1] https://www.ctan.org/tex-archive/macros/latex/contrib/hyperref?lang=en